39580284: add new RPC command z_viewtransaction to get detailed shielded information about in-wallet transaction <txid>
pasteld v1.1.2
Added new RPC command 'z_viewtransaction' get detailed shielded information about in-wallet transaction <txid>:
Arguments:
1. "txid"    (string, required) The transaction id

Result:
{
  "txid" : "transactionid",   (string) The transaction id
  "spends" : [
    {
      "type" : "sprout|sapling",      (string) The type of address
      "js" : n,                       (numeric, sprout) the index of the JSDescription within vJoinSplit
      "jsSpend" : n,                  (numeric, sprout) the index of the spend within the JSDescription
      "spend" : n,                    (numeric, sapling) the index of the spend within vShieldedSpend
      "txidPrev" : "transactionid",   (string) The id for the transaction this note was created in
      "jsPrev" : n,                   (numeric, sprout) the index of the JSDescription within vJoinSplit
      "jsOutputPrev" : n,             (numeric, sprout) the index of the output within the JSDescription
      "outputPrev" : n,               (numeric, sapling) the index of the output within the vShieldedOutput
      "address" : "zcashaddress",     (string) The Zcash address involved in the transaction
      "value" : x.xxx                 (numeric) The amount in )" + CURRENCY_UNIT + R"(
      "valuePsl" : xxxx               (numeric) The amount in patoshis
    }
    ,...
  ],
  "outputs" : [
    {
      "type" : "sprout|sapling",      (string) The type of address
      "js" : n,                       (numeric, sprout) the index of the JSDescription within vJoinSplit
      "jsOutput" : n,                 (numeric, sprout) the index of the output within the JSDescription
      "output" : n,                   (numeric, sapling) the index of the output within the vShieldedOutput
      "address" : "zcashaddress",     (string) The Zcash address involved in the transaction
      "outgoing" : true|false         (boolean, sapling) True if the output is not for an address in the wallet
      "value" : x.xxx                 (numeric) The amount in )" + CURRENCY_UNIT + R"(
      "valuePsl" : xxxx               (numeric) The amount in patoshis
      "memo" : "hexmemo",             (string) Hexademical string representation of the memo field
      "memoStr" : "memo",             (string) Only returned if memo contains valid UTF-8 text.
    }
    ,...
  ],
}

This command will decrypt the transaction with all Sapling keys in the local wallet and return the shielded information that
it can recover, including:
  - the address that each note belongs to;
  - values in both decimal PSL and patochis
  - the id of the transaction that each spent note was received in ;
  - an outgoing flag on each new note, which will be true if the output is not for an address in the wallet;
  - a memoStr field for each new note, containing it text memo (if its memo field contains a valid utf-8 string)

Added new RPC command 'z_getnotescount':
Returns the number of sprout and sapling notes available in the wallet

Arguments:
1. minconf      (numeric, optional, default=1) Only include notes in transactions confirmed at least this many times.

Result:
{
  "sprout"      (numeric) the number of sprout notes in the wallet
  "sapling"     (numeric) the number of sapling notes in the wallet
}

Visual Studio 2019 projects:
  - added new project bitcoin_test.vcxproj for src/test - bitcoin tests
  - added ENABLE_MINING option, ported equihash solver to Visual C++

3rdparty libraries:
  - upgraded nlohmann json parser from v3.2.0 (08/20/2018) to v3.9.1 (08/06/2020) (https://github.com/nlohmann/json/)
  - upgraded univalue (High performance RAII C++ JSON library and universal value object class) from v1.0.3 (01/29/2017) to
    v1.1.1 (11/11/2019) (https://github.com/jgarzik/univalue)
  - added utfcpp library v3.2 (05/03/2021) to src/utfcpp for portable utf-8 support for c++17 (https://github.com/nemtrif/utfcpp)
    this can be removed when we switch to c++20

libbitcoin_common changes:
  - Base58Type, Bech32Type constants moved to separate class KeyConstants, CChainParams was inheritied from that class
  - KeyConstants introduces two new methods Base58Prefix, Bech32HRP that returns base58 prefixes and bech32-address by type for specific net-type
  - introduced new bech32 type - sapling extended full viewing key (SAPLING_EXTENDED_FVK):
       mainnet bech32-address prefix: pxviews
       testnet bech32-address prefix: pxviewtestsapling
       regtest bech32-address prefix: pxviewregtestsapling
  -key_io methods are encapsulated into KeyIO class with KeyConstants member initialized for the specific net-type

src/zcash changes:
   - zip32.h/cpp moved to zcash/address folder, sapling/sprout keys handling was moved from Address.hpp/cpp to corresponding files address/sapling.hpp/cpp, address/sprout.hpp/cpp
   - <boost/variant.hpp> was replaced with std::variant, new std::variant used to process encoding for sprout/sapling payment addresses/viewing keys/spending keys
   - SaplingFullViewingKey replaced with SaplingExtendedFullViewingKey

libbitcoin_wallet changes:
   - atomic update counter for wallet db
   - added support for writing to wallet db sapling extended full viewing key
   - sprout/sapling keys handling with std::visit (replaced boost::static_visitor)
   - replaced all push_back(Pair()) for univalues with more effective pushKV that supports C++11 move semantics
   - paymentdisclosure.cpp|h & paymentdisclosuredb.cpp|h moved to src/wallet

build changes/rebranding:
   - renamed bitcoind.cpp to pasteld.cpp, bitcoinfd-res.rc to pasteld-res.rc
   - changed autoconf values in configure.ac:
	build bitcoind -> build_pasteld,
	build_bitcoin_utils -> build_pastel_utils,
	bitcoin_daemon_name -> pastel_deamon_name,
	bitcoin_cli_name -> pastel_cli_name,
	bitcoin_tx_name -> pastel_tx_name
   - added AC_CANONICAL_TARGET to remove configure warning
   - changed default enable_mining option to yes (no tests are working without mining support)

libbitcoin_crypto:
   - implemented equihash solver on Visual C++ (pow/tromp/equi_miner.h), this enables rpc debugging on Visual Studio
   - added implementation of pthread_barrier for VC c++11 (https://github.com/nipuntalukdar/cyclicbarrier), used in equihash solver

tests:
   - added tests for z_viewtransaction to qa/rpc-tests/wallet_listreceived.py
   - added tests (src/gtest/test_keys.cpp) for sapling extended full viewing key

+++ src/base58.cpp
@@ -33,7 +33,7 @@ bool DecodeBase58(const char* psz, std::vector<unsigned char>& vch)
         if (ch == NULL)
             return false;
         // Apply "b256 = b256 * 58 + ch".
-        int carry = ch - pszBase58;
+        size_t carry = ch - pszBase58;
         for (std::vector<unsigned char>::reverse_iterator it = b256.rbegin(); it != b256.rend(); it++) {
             carry += 58 * (*it);
             *it = carry % 256;
